Refactors ApifyPipeline module architecture
Reorganizes the ApifyPipeline into a more structured and maintainable architecture.

Moves components into clearly defined layers to enhance separation of concerns:
- `Core`: Contains core domain models and transformations.
- `DataAccess`: Manages persistence logic, including repositories, queries, and migrations.
- `ExternalServices`: Encapsulates integrations with third-party APIs like Apify and Supabase.
- `Infrastructure`: Houses utilities and environmental configurations.
- `Web/Application/Commands`: Dedicated to web API endpoints and their command handlers, replacing previous `Scheduler` based execution.
- `Background/Jobs`: Renames `Scheduler/Actors` to `Background/Jobs` for better descriptive clarity of background tasks.

This refactoring improves modularity, clarity, and overall maintainability of the codebase.
