+++ src/ApifyPipeline/Background/Jobs/TweetCollector/index.ts
@@ -1,0 +1,1 @@
+import './TweetCollectorJob';
