+++ src/ApifyPipeline/DataAccess/Queries/queryBuilder.ts
