--- src/ApifyPipeline/Scheduler/Actors/TweetCollector/index.ts
@@ -1,1 +1,0 @@
-import './TweetCollectorActor';
