--- src/ApifyPipeline/Scheduler/StartApifyRunEndpoint/index.ts
@@ -1,1 +1,0 @@
-export { startApifyRunEndpoint } from './StartApifyRunEndpoint';
