--- src/ApifyPipeline/Scheduler/StartApifyRunExecutor/index.ts
@@ -1,6 +1,0 @@
-export {
-  startApifyRunExecutor,
-  type StartApifyRunExecutor as StartApifyRunCommand,
-  type StartApifyRunExecutorInput as StartApifyRunCommandInput,
-  type StartApifyRunExecutorResult as StartApifyRunCommandResult,
-} from './StartApifyRunExecutor';
