--- src/ApifyPipeline/Ui/Application/Domain/Utilities/chunk.ts
@@ -1,13 +1,0 @@
-export const chunkArray = <T>(items: T[], chunkSize: number): T[][] => {
-  if (chunkSize <= 0) {
-    throw new Error('chunkSize must be greater than zero.');
-  }
-
-  const result: T[][] = [];
-
-  for (let index = 0; index < items.length; index += chunkSize) {
-    result.push(items.slice(index, index + chunkSize));
-  }
-
-  return result;
-};
