+++ src/ApifyPipeline/Web/Application/Commands/StartApifyRun/index.ts
@@ -1,0 +1,9 @@
+export {
+  commandSchema,
+  type StartApifyRunCommand,
+  type StartApifyRunCommandInput,
+} from './StartApifyRunCommand';
+export {
+  startApifyRunCommandHandler,
+  type StartApifyRunCommandResult,
+} from './StartApifyRunCommandHandler';
